+++ tests/sdk/v2/driver/test_ce_runtime.py
@@ -2,7 +2,7 @@
 ################################################################################
 from datetime import timedelta
 from pathlib import Path
-from unittest.mock import DEFAULT, MagicMock, Mock, call
+from unittest.mock import DEFAULT, MagicMock, Mock, call, create_autospec
 
 import pytest
 
@@ -14,8 +14,14 @@ from orquestra.sdk._base._testing._example_wfs import (
     workflow_with_different_resources,
 )
 from orquestra.sdk.schema.configs import RuntimeConfiguration, RuntimeName
+from orquestra.sdk.schema.ir import WorkflowDef
 from orquestra.sdk.schema.responses import JSONResult
-from orquestra.sdk.schema.workflow_run import State, WorkflowRunId
+from orquestra.sdk.schema.workflow_run import (
+    RunStatus,
+    State,
+    WorkflowRun,
+    WorkflowRunId,
+)
 
 
 @pytest.fixture
@@ -96,6 +102,24 @@ def workflow_def_id():
 @pytest.fixture
 def workflow_run_id():
     return "00000000-0000-0000-0000-000000000000"
+
+
+@pytest.fixture
+def workflow_run_status(workflow_run_id: WorkflowRunId):
+    def _workflow_run(state: State):
+        workflow_def_mock = create_autospec(WorkflowDef)
+        return WorkflowRun(
+            id=workflow_run_id,
+            workflow_def=workflow_def_mock,
+            task_runs=[],
+            status=RunStatus(
+                state=state,
+                start_time=None,
+                end_time=None,
+            ),
+        )
+
+    return _workflow_run
 
 
 class TestCreateWorkflowRun:
@@ -481,17 +505,63 @@ class TestGetWorkflowRunResultsNonBlocking:
             with pytest.raises(exceptions.WorkflowRunNotFoundError):
                 _ = runtime.get_workflow_run_outputs_non_blocking(workflow_run_id)
 
-        def test_no_results(
+        def test_no_results_not_succeeded(
             self,
             mocked_client: MagicMock,
             runtime: _ce_runtime.CERuntime,
             workflow_run_id: str,
+            workflow_run_status,
         ):
             # Given
             mocked_client.get_workflow_run_results.return_value = []
+            mocked_client.get_workflow_run.return_value = workflow_run_status(
+                State.RUNNING
+            )
             # When
             with pytest.raises(exceptions.WorkflowRunNotSucceeded):
                 _ = runtime.get_workflow_run_outputs_non_blocking(workflow_run_id)
+
+        def test_no_results_succeeded(
+            self,
+            monkeypatch: pytest.MonkeyPatch,
+            mocked_client: MagicMock,
+            runtime: _ce_runtime.CERuntime,
+            workflow_run_id: str,
+            workflow_run_status,
+        ):
+            # Given
+            mocked_client.get_workflow_run_results.return_value = []
+            mocked_client.get_workflow_run.return_value = workflow_run_status(
+                State.SUCCEEDED
+            )
+            monkeypatch.setattr(_ce_runtime._retry.time, "sleep", Mock())
+            # When
+            with pytest.raises(exceptions.WorkflowResultsNotReadyError):
+                _ = runtime.get_workflow_run_outputs_non_blocking(workflow_run_id)
+
+            # We should try a few times if the results were not ready
+            assert mocked_client.get_workflow_run_results.call_count == 5
+
+        def test_eventually_get_results(
+            self,
+            monkeypatch: pytest.MonkeyPatch,
+            mocked_client: MagicMock,
+            runtime: _ce_runtime.CERuntime,
+            workflow_run_id: str,
+            workflow_run_status,
+        ):
+            # Given
+            mocked_client.get_workflow_run_results.side_effect = [[], [], [Mock()]]
+            mocked_client.get_workflow_run.return_value = workflow_run_status(
+                State.SUCCEEDED
+            )
+            monkeypatch.setattr(_ce_runtime.serde, "deserialize", lambda x: x)
+            monkeypatch.setattr(_ce_runtime._retry.time, "sleep", Mock())
+            # When
+            _ = runtime.get_workflow_run_outputs_non_blocking(workflow_run_id)
+
+            # We should have the results after 3 attempts
+            assert mocked_client.get_workflow_run_results.call_count == 3
 
         def test_unknown_http(
             self,