+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Setting workflow_id and project_id is now available on workflow Python API start() and prepare() functions
 
 🐛 *Bug Fixes*
+* Retry getting results from CE if the results were not ready but the workflow succeeded.
 
 💅 *Improvements*